requirements.txt
numpy>=1.21.0
gymnasium>=0.28.0
termcolor>=2.0.0
scipy>=1.7.0
torch>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
matplotlib>=3.5.0
imageio>=2.9.0
Pillow>=9.0.0
tqdm>=4.64.0
pytest>=7.0.0
wandb>=0.15.0
joblib>=1.0.0
seaborn>=0.11.0
click>=8.0.0
rich>=10.0.0

